Storage of arguments, input of operators
Pressing an operator button will store the current value as a variable to be used as an argument when performing calculator operations as well as storing which operation is to be performed.
When two arguments already exist to be operated upon, the existing two are combined according to the previous operation selected and stored in the secondValue variable. The new operator is then stored in place of the old and additional arguments can be used.